pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optirag"
version = "0.1.0"
description = "Automated RAG configuration optimization (Optuna + RAGAS; Stage 2 Deep Agents)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "google-genai>=1.0",
    "pinecone>=5.0",
    "optuna>=3.6",
    "ragas>=0.2",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.3",
    "datasets>=3.0",
    "beir>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "numpy>=1.26",
]

[project.optional-dependencies]
agents = [
    "deepagents>=0.1.0",
    "langgraph>=0.2",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
optirag = "optirag.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
markers = [
    "integration: requires API keys / network",
]

[tool.mypy]
python_version = "3.11"
packages = ["optirag"]
strict = false
